test(refactor): 테스트 정리
변경 내용
- 'given~~' 형태의 변수명에서 'given'을 제거
- 오타 수정
- 15줄 이하인지 체크
- 불필요한 테스트 데이터 생성기 제거
- 테스트 클래스 위치 변경

변경 이유
- 규칙에 맞게 테스트를 작성하기 위해 수정
- 가독성을 높이기 위해 수정
